requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
pymongo==4.6.1
